+++ packages/perseus-editor/src/widgets/image-editor/components/image-settings.tsx
@@ -1,15 +1,15 @@
 import {components} from "@khanacademy/perseus";
 import {isFeatureOn} from "@khanacademy/perseus-core";
 import {LabeledField} from "@khanacademy/wonder-blocks-labeled-field";
-import {sizing} from "@khanacademy/wonder-blocks-tokens";
-import {HeadingXSmall} from "@khanacademy/wonder-blocks-typography";
 import * as React from "react";
 
-import {AutoResizingTextArea} from "../../components/auto-resizing-text-area";
+import {AutoResizingTextArea} from "../../../components/auto-resizing-text-area";
+import {wbFieldStyles, wbFieldStylesWithDescription} from "../utils";
 
-import styles from "./image-editor.module.css";
+import DecorativeToggle from "./decorative-toggle";
+import ImageDimensionsInput from "./image-dimensions-input";
 
-import type {Props} from "./image-editor";
+import type {Props} from "../image-editor";
 
 const {SvgImage} = components;
 
@@ -25,6 +25,7 @@ export default function ImageSettings({
     backgroundImage,
     apiOptions,
     caption,
+    decorative,
     longDescription,
     title,
     onChange,
@@ -34,15 +35,17 @@ export default function ImageSettings({
         null,
     );
 
-    if (!backgroundImage.url) {
+    if (
+        !backgroundImage.url ||
+        !backgroundImage.width ||
+        !backgroundImage.height
+    ) {
         return null;
     }
 
-    const dimensions = `${backgroundImage.width} x ${backgroundImage.height}`;
-    const dimensionString =
-        backgroundImage.width && backgroundImage.height
-            ? dimensions
-            : "unknown";
+    const hasPopulatedFields = Boolean(
+        alt || caption || title || longDescription,
+    );
 
     // Show "alt text too long" error on change so the user is notified
     // as they type that they're writing too much.
@@ -82,20 +85,19 @@ export default function ImageSettings({
             />
 
             {/* Dimensions */}
-            <div className={styles.dimensionsContainer}>
-                <HeadingXSmall
-                    style={{
-                        // TODO: Use CSS modules after Wonder Blocks styles
-                        // are moved to a different layer.
-                        padding: 0, // reset default padding
-                        marginInlineEnd: sizing.size_080,
-                        color: "var(--wb-semanticColor-core-foreground-neutral-strong)",
-                    }}
-                >
-                    Dimensions:
-                </HeadingXSmall>
-                {dimensionString}
-            </div>
+            <ImageDimensionsInput
+                backgroundImage={backgroundImage}
+                onChange={onChange}
+            />
+
+            {/* Decorative */}
+            {imageUpgradeFF && (
+                <DecorativeToggle
+                    decorative={decorative}
+                    hasPopulatedFields={hasPopulatedFields}
+                    onChange={onChange}
+                />
+            )}
 
             {/* Alt text */}
             <LabeledField
@@ -106,6 +108,7 @@ export default function ImageSettings({
                         value={alt ?? ""}
                         onBlur={(e) => handleAltFieldBlur(e.target.value)}
                         onChange={handleAltFieldChange}
+                        disabled={decorative}
                     />
                 }
                 errorMessage={altFieldError}
@@ -122,6 +125,7 @@ export default function ImageSettings({
                             onChange={(value) =>
                                 onChange({longDescription: value})
                             }
+                            disabled={decorative}
                         />
                     }
                     styles={wbFieldStyles}
@@ -135,6 +139,7 @@ export default function ImageSettings({
                     <AutoResizingTextArea
                         value={title ?? ""}
                         onChange={(value) => onChange({title: value})}
+                        disabled={decorative}
                     />
                 }
                 styles={wbFieldStyles}
@@ -147,6 +152,7 @@ export default function ImageSettings({
                     <AutoResizingTextArea
                         value={caption ?? ""}
                         onChange={(value) => onChange({caption: value})}
+                        disabled={decorative}
                     />
                 }
                 styles={wbFieldStyles}
@@ -154,28 +160,3 @@ export default function ImageSettings({
         </>
     );
 }
-
-// TODO: Use CSS modules after Wonder Blocks styles
-// are moved to a different layer.
-const wbFieldStyles = {
-    root: {
-        marginBlockEnd: sizing.size_080,
-    },
-    label: {
-        fontWeight: "bold",
-        paddingBlockEnd: sizing.size_040,
-    },
-};
-
-// Exporting so this can be used in image-url-input.tsx.
-export const wbFieldStylesWithDescription = {
-    ...wbFieldStyles,
-    label: {
-        ...wbFieldStyles.label,
-        paddingBlockEnd: 0,
-    },
-    description: {
-        paddingBlockStart: 0,
-        paddingBlockEnd: sizing.size_040,
-    },
-};