+++ packages/perseus-core/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @khanacademy/perseus-core
 
+## 19.5.0
+
+### Minor Changes
+
+-   [#2946](https://github.com/Khan/perseus/pull/2946) [`658b772c88`](https://github.com/Khan/perseus/commit/658b772c88f519b15006962bf5f27b91711e57c6) Thanks [@SonicScrewdriver](https://github.com/SonicScrewdriver)! - Update the Radio Widget current version to 3.0
+
+## 19.4.0
+
+### Minor Changes
+
+-   [#2918](https://github.com/Khan/perseus/pull/2918) [`42c456d99e`](https://github.com/Khan/perseus/commit/42c456d99e0002589c088c721f09cbb26ea57b2c) Thanks [@ivyolamit](https://github.com/ivyolamit)! - Add new decorative props in Image Widget
+
 ## 19.3.0
 
 ### Minor Changes